requirements.txt
numpy
pandas
opencv-python
Pillow
matplotlib
torch
torchvision
tensorflow
keras
scikit-learn
scipy
lxml
beautifulsoup4
